refactor: split ArticlePage.tsx into focused sub-components (#834)
Extracts renderMarkdown, TTS audio, keyboard shortcuts, swipe nav,
sharing/highlights state, and the reader UI (header, meta, body, action
bar, insights, perspectives, why-recommended) out of the 857-line
ArticlePage.tsx into dedicated hooks/lib/components. No behavior,
routes, or keyboard bindings change; ArticlePage.tsx now composes the
extracted pieces and is under 300 lines.
